+++ .gitignore
@@ -104,4 +104,5 @@ venv.bak/
 .mypy_cache/
 
 #config files
-server/api/resources/api42config.py
+server/rq42/api42config.py
+server/api/config/dbConfig.py
