+++ server/api/models/user.py
@@ -1,3 +1,4 @@
+from flask_marshmallow import Marshmallow
 from datetime import datetime
 from api.db import db
 
@@ -6,7 +7,7 @@ class User(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     id_user42 = db.Column(db.Integer, nullable=False)
-    login = db.Column(db.String(45), nullable=False)
+    login = db.Column(db.String(45), unique=True, nullable=False)
 
     #   relationship with 'Mentors' table, Mentor Model Class
     mentors = db.relationship('Mentor', backref='user', lazy=True)
@@ -24,3 +25,4 @@ class User(db.Model):
             'id': self.id,
             'id_user42': self.id_user42,
             'login': self.login}
+
