+++ server/api/resources/__init__.py
@@ -1,36 +1,4 @@
-import threading, time
-from api.config import apikeys
-from .api42 import *
-
-api42Requester = Api42(apikeys.uid, apikeys.secret)
-
-class globalOnlineUsers():
-
-    _onlineUsersList = []
-    onlineMentorLock = threading.Lock() 
-
-    @staticmethod
-    def updateOnlineUsers():
-        while True:
-            globalOnlineUsers.lock()
-            globalOnlineUsers._onlineUsersList = api42Requester.onlineStudents()
-            globalOnlineUsers.unlock()
-            time.sleep(120)
-    
-    #   Locks access to the _onlineMentorList
-    @staticmethod
-    def lock():
-        globalOnlineUsers.onlineMentorLock.acquire()
-    
-    #   Unlocks access to the _onlineMentorList    
-    @staticmethod
-    def unlock():
-        globalOnlineUsers.onlineMentorLock.release()
-
-    @staticmethod
-    def run():
-        threading.Thread(target=globalOnlineUsers.updateOnlineUsers).start()
-
+import api42
 from .users import *
 from .projects import *
 from .appointments import *
